requirements.txt
pillow==8.1.0
colorama==0.4.4